refactor: Move OmniBaseError to exceptions.py and update all imports - separates Python exception from Pydantic models
