+++ .onextree
@@ -62,8 +62,8 @@ children:
       type: file
     - name: test_registry.py
       type: file
-  - name: errors.py
-    type: file
+- name: exceptions.py
+  type: file
 - name: fixtures
   type: directory
   children:
